+++ packages/atproto/lib/atproto.dart
@@ -60,6 +60,8 @@ export 'package:atproto/src/services/entities/subscribed_repo_migrate.dart';
 export 'package:atproto/src/services/entities/subscribed_repo_tombstone.dart';
 export 'package:atproto/src/services/entities/update_action.dart';
 export 'package:atproto/src/services/entities/signing_key.dart';
+export 'package:atproto/src/services/entities/signup_queue.dart';
+export 'package:atproto/src/services/entities/transferred_account.dart';
 export 'package:atproto/src/services/identity_service.dart';
 export 'package:atproto/src/services/constants/moderation_reason_type.dart';
 export 'package:atproto/src/services/moderation_service.dart';
@@ -67,6 +69,7 @@ export 'package:atproto/src/services/repo_service.dart';
 export 'package:atproto/src/services/server_service.dart';
 export 'package:atproto/src/services/sync_service.dart';
 export 'package:atproto/src/services/label_service.dart';
+export 'package:atproto/src/services/temp_service.dart';
 export 'package:atproto_core/atproto_core.dart'
     show
         XRPCResponse,